refactor(PT-74): restructure validation, clean up and reorganize files
