+++ index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -7,10 +7,10 @@
     <title>Pet Help</title>
   </head>
   <body>
-     <div style="display: none;">
+    <div style="display: none">
       <object type="image/svg+xml" data="/icons.svg"></object>
     </div>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/app/main.tsx"></script>
   </body>
 </html>
