+++ src/app/main.tsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import '@/index.css';
-import App from '@/components/App/App.tsx';
+import App from '@/app/App';
 import { Provider } from 'react-redux';
-import { store } from '@/store/store';
+import { store } from '@/app/store';
 import { BrowserRouter } from 'react-router-dom';
 
 const rootElement = document.getElementById('root')!;
