+++ src/features/auth/validations/loginSchema.ts
@@ -1,0 +1,9 @@
+import { emailField, passwordField } from '@/shared/validations/fields';
+import { z } from 'zod';
+
+export const loginSchema = z.object({
+  email: emailField,
+  password: passwordField,
+});
+
+export type LoginSchemaType = z.infer<typeof loginSchema>;
