+++ src/features/booking/validation/bookingSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { emailRegex, streetOrHouseRegex } from '@/shared/constants/regex';
-import { timeSlots } from '@/shared/constants/bookingTime';
+import { timeSlots } from '@/features/booking/constants/bookingTime';
 import { dogWeightRanges } from '@/shared/constants/dogsWeight';
 
 const baseSchema = z.object({
