+++ src/features/hero/heroSelectors.ts
@@ -1,4 +1,4 @@
 // src/features/hero/heroSelectors.ts
-import { RootState } from "@/store/store";
+import { RootState } from '@/app/store';
 
 export const selectAnimal = (state: RootState) => state.hero.animal;
