+++ src/features/review/validation/reviewSchema.ts
@@ -1,4 +1,8 @@
-import { animalField, nameField, serviceField } from '@/validations/fields';
+import {
+  animalField,
+  nameField,
+  serviceField,
+} from '@/shared/validations/fields';
 import { z } from 'zod';
 
 export const reviewSchema = z
