+++ src/shared/components/UI/Modal.tsx
@@ -56,7 +56,7 @@ const Modal: FC<ModalProps> = ({
     >
       <div
         className={clsx(
-          'relative bg-alabaster rounded-3xl xl:rounded-2xl w-[344px] xl:w-[496px] py-16 xl:py-10 px-5 xl:px-24 transition-all duration-500 ease-in-out transform',
+          'relative bg-alabaster rounded-3xl xl:rounded-2xl w-[345px] xl:w-[496px] py-16 xl:py-10 px-5 xl:px-24 transition-all duration-500 ease-in-out transform',
           isOpen
             ? 'opacity-100 scale-100 translate-y-0'
             : 'opacity-0 scale-95 -translate-y-4',
