+++ src/shared/constants/animals.ts
@@ -1,0 +1,3 @@
+// shared/constants/animals.ts
+export const animals = ['собаки', 'коти'] as const;
+export type Animal = (typeof animals)[number];
